HTTP関連を別ファイル化 , 不要な関数とコメントを消した
now関数がGetDateと同じような挙動だったのでGetDateを使うようにした

HTTP関連のところを関数化して別ファイルとした
