+++ util/doorHttp.js
@@ -12,8 +12,6 @@ process.on('exit', function () {
 
 var fs = require('fs'),
     color = require('../src/color'),
-    //http = require('http'),
-    //server = {},
     IRMagician = require('../irMagician'),
     irMagician = {},
     port = '',
@@ -28,7 +26,9 @@ var fs = require('fs'),
     mailer,
     mailerPass,
 
-    GetDate = require('../src/getDate.js'),
+    getPost = require('../src/GetPost'),
+
+    GetDate = require('../src/getDate'),
     getDate = new GetDate();
 
 
@@ -49,7 +49,6 @@ lightJudge = function(){
     lightJudge.closeCount = 0;
   }
   lightJudge.closeCount++;
-  //console.log('count' + lightJudge.closeCount);
   if(lightJudge.closeCount % 2 === 1){
     dataName = '../json/lightOn.json';
   }else{
@@ -57,7 +56,6 @@ lightJudge = function(){
   }
   irMagician.Lplay(dataName, function(){console.log('Lplay end callback');});
 };
-//lightJudge.closeCount = 0;
 
 arduino.on('open', function(){
     console.log(color.info('arduino is opened'));
@@ -84,103 +82,5 @@ arduino.on('open', function(){
     });
 });
 //TODO:メール送信しなくていい時の判定追加
-
-
-
-var http = require('http'),
-  server = {};
-
-server = http.createServer(function(req, res){
-  switch (req.method){
-    case 'POST':
-      var data = '';
-      res.writeHead(200,{'Access-Control-Allow-Origin':"*"});
-      req.setEncoding('utf8');
-      req.on('data', function(dataChunk) {
-        // データ受信中の処理
-        data += dataChunk;
-      });
-      req.on('end', function() {
-        // データ受信完了後の処理
-        console.log('posted data : ' + data);
-        onPost(data);//dataの処理
-        res.end(data+'');
-      });
-      break;
-
-    case 'GET':
-      console.log('req: '+req.url);
-      if(req.url === '/on'){
-        send('ON');
-        res.writeHead(200,{'Content-Type':'text/html'});
-        res.end('<h1>ON</h1>');
-        break;
-      }else{
-        if(req.url === '/off'){
-          send('OFF');
-          res.writeHead(200,{'Content-Type':'text/html'});
-          res.end('<h1>OFF</h1>');
-          break;
-        }else{
-          res.writeHead(200,{'Content-Type':'text/html'});
-          res.end('<h1>ON? OFF?</h1>');
-          break;
-        }
-      }
-  }
-});
-port = 8080;
-server.listen(port);
-console.log('lidten at '+port);
-
-function onPost(data){
-  switch (data+''){
-    case 'lightOFF': send('OFF');break;
-    case 'lightON' : send('ON');break;
-    default        : console.log('data is '+data);
-  }
-}
-
-function send(judge){
-  var word = '',
-      playDataName = '';
-
-  if(judge==='ON'){
-    word = judge;
-    playDataName = '../json/lightOn.json';
-  }else{
-    if(judge==='OFF'){
-      word = 'OFF';
-      playDataName = '../json/lightOff.json';
-    }
-  }
-
-  console.log(color.safe('<'+now()+'> light'+word));
-  irMagician.Lplay(playDataName, function(){console.log('Lplay end callback');});
-}
-
-function now(){
-  var getTime = new Date(),
-      month = getTime.getMonth()+1,
-      date = getTime.getDate(),
-      hours = getTime.getHours(),
-      minutes = getTime.getMinutes(),
-      seconds = getTime.getSeconds();
-  return dateFormatter([month,date,hours,minutes,seconds]);
-}
-
-function dateFormatter(array){
-  var res = '';
-  if(array.length !== 5){
-    throw 'dateFormatterの引数がまちがってる';
-  }
-  for(var i=0;i<array.length;i++){
-    if(array[i] < 10){
-      array[i] = '0'+array[i];
-    }
-  }
-  res = array[0]+'/'+array[1]+' '+array[2]+':'+array[3]+':'+array[4];
-  return res;
-}
 //FIXME:lightOffの時だけ応答ない時あり
 //TODO:扉のセンサで、ライト点けたり消したりしなくていい時あるのでその判定
